move db logic to dbOps
